requirements.txt
CTkMessagebox==2.7
click==8.1.8
colorama==0.4.6
customtkinter==5.2.2
cvzone==1.6.1
darkdetect==0.8.0
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
future==1.0.0
mysql-connector-python==9.2.0
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pip==24.2
setuptools==78.1.0
wheel==0.45.1